requirements.txt
# Kearney AI Commons - Python Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
Flask>=3.0.0

# Environment Management
python-dotenv>=1.0.0

# AI Provider SDKs
anthropic>=0.40.0
openai>=1.50.0

# Data Processing
pandas>=2.0.0

# Development Tools (optional)
pytest>=8.0.0
black>=24.0.0
